framework-comparisons/04-memory-management/langchain/requirements.txt
langchain==1.2.6
langchain-openai==1.1.7
python-dotenv==1.2.1
numpy<2.0.0
